+++ packages/components/ui/src/icons/IconMenu.svelte
@@ -14,7 +14,9 @@
 	height={size}
 	width={size}
 >
-	<polyline points="15 18 9 12 15 6" />
+	<line class="top-bun" x1="3" y1="12" x2="21" y2="12" />
+	<line class="meat" x1="3" y1="6" x2="21" y2="6" />
+	<line class="bottom-bun" x1="3" y1="18" x2="21" y2="18" />
 </Icon>
 
-<!-- feathericons/feather/icons/chevrons-left.svg -->
+<!-- feathericons/feather/icons/menu.svg -->